app/requirements.txt
celery==5.3.4
redis==5.0.1
paramiko==3.4.0
python-dotenv==1.0.0
flower==2.0.1
